No more integer dt
Sorry Mario :(

+++ src/scene_manager.h
@@ -5,7 +5,7 @@
 struct Scene {
 	virtual void EnterScene() = 0;
 	virtual void ExitScene() = 0;
-	virtual void Update(int dt) = 0;
+	virtual void Update(float dt) = 0;
 	virtual void Draw(sf::RenderTarget& window, bool debug) = 0;
 	virtual ~Scene() { }
 };
